refactor(quality): reduce audioPlayer.ts complexity by splitting into focused modules
- Extract Discord gateway bridge into src/voiceAdapter.ts (buildAdapter + adapter cleanup state)
- Extract SFX file playback into src/sfxPlayer.ts (playFile, path validation, ffmpeg init)
- Simplify isPermanentMisconfigurationError to a single boolean expression grouped by category
- Extract connection event handlers (handleConnectionError, handleDisconnected) to reduce
  setupConnectionHandlers and connect() cyclomatic complexity
- Extract tearDownPlayer, releasePreviousConnection, cleanupFailedConnect helpers
- Add isConnected() and startPlayback() bridge exports to audioPlayer.ts for sfxPlayer.ts
- Update commandRouter.ts to import playFile from sfxPlayer

File complexity reduced from 87 → 54 across three focused modules with no circular imports.
